+++ src/main/java/com/tobeto/rentacar/services/concretes/EmployeeManager.java
@@ -57,7 +57,7 @@ public class EmployeeManager implements EmployeeService {
         Employee employee =this.modelMapperService.forRequest()
                 .map(updateEmployeeRequests, Employee.class);
         employee.setId(updateEmployeeRequests.getId());
-        employee.setSalary(updateEmployeeRequests.getSalary());
+        employee.setSalary((float) updateEmployeeRequests.getSalary());
 
         this.employeeRepository.save(employee);
